feat(ff-filter): renamed filter_name module to ffmpeg_token and moved to directory mod.rs
