+++ crates/ff-filter/src/graph/mod.rs
@@ -2,7 +2,7 @@
 
 pub mod builder;
 pub mod composition;
-mod filter_name;
+mod ffmpeg_token;
 pub(crate) mod filter_step;
 #[allow(clippy::module_inception)]
 mod graph;
@@ -13,7 +13,7 @@ pub use composition::{
     AudioConcatenator, AudioTrack, ClipJoiner, ClipTransition, MultiTrackAudioMixer,
     MultiTrackComposer, ProxySource, VideoConcatenator, VideoLayer,
 };
-pub use filter_name::FfmpegToken;
+pub use ffmpeg_token::FfmpegToken;
 pub use filter_step::FilterStep;
 pub use graph::FilterGraph;
 pub use types::{
